Major: Added socket-based multirun capability.  See "Multirun mode" in the GLD Wiki documentation for more details. Added '--slave' and '--slavenode' command line arguments.  Slavenod starts a tight loop for GridLAB-D and opens a listener socket for remote instances that will use socket-based multirun mode. Minor: Fixed warning when downcasting an int64 keys->value to a uint32 for convert_to_enumeration.  Enumerations are not likely to have more than 2^30 values. Added "id" to the global variables.  For internal use with socket-based multirun. Added slavenode processing to exec.c. Added disabled code for 'OR' keyword in find.c.  Is not logically correct. Split 'instance' into 'instance_cnx' and 'instance_slave' as well, for organization. Added odd null checks, cleaned out superfluous comments in instance and linkage files. Hostnames now support colons, to include port numbers. Instance blocks now include "execdir" and "return_port", which specify where to find the executable for a given slave run, and which port on the master that the slave should connect back to (auto-assigned if blank). Fixed a number of odd casting warnings in various files.
git-svn-id: https://svn.code.sf.net/p/gridlab-d/code/trunk@2948 54b63ff4-7b51-0410-8bda-9fa96797e505

+++ core/convert.cpp
@@ -268,7 +268,7 @@ int convert_to_enumeration(char *buffer, /**< a pointer to the string buffer */
 	{
 		if (strcmp(keys->name,buffer)==0)
 		{
-			*(uint32*)data=keys->value;
+			*(uint32*)data=(uint32)(keys->value);
 			found = true;
 			break;
 		}
